Consolidate all Activity resolver logic to ActivitiesResolver, fix cache
