+++ packages/lib/urql/updates/index.ts
@@ -2,7 +2,6 @@ export * from './acter-connection-delete'
 export * from './acter-create'
 export * from './acter-delete'
 export * from './activity-create'
-export * from './activity-update'
 export * from './link-create'
 export * from './link-delete'
 export * from './post-create'
